Update unique_domains.r
This R script takes a domain annotation file with possible duplicate domain names, renames them to be unique, and writes a cleaned version with the new names (e.g. a transcript with multiple CIDRα1.1s have a suffix added to them such that it is now called CIDRα1.1, CIDRα1.1.1, CIDRα1.1.2). This script is called in the Size filter and N50 values script.